requirements.txt
# Data manipulation and analysis
pandas>=1.5.3,<2.0.0
numpy>=1.24.3,<2.0.0

# Data visualization
matplotlib>=3.7.1,<4.0.0
seaborn>=0.12.2,<0.13.0
plotly>=5.14.1,<6.0.0

# Machine learning
scikit-learn>=1.2.2,<2.0.0
statsmodels>=0.14.0,<0.15.0

# Deep learning
tensorflow>=2.13.0,<3.0.0
torch>=2.0.1,<3.0.0

# Natural Language Processing
nltk>=3.8.1,<4.0.0
spacy>=3.5.3,<4.0.0

# Data collection
yfinance>=0.2.18,<0.3.0
pandas-datareader>=0.10.0,<0.11.0
requests>=2.31.0,<3.0.0

# Technical analysis
pandas-ta>=0.3.14b0,<0.4.0

# Web framework (optional, for API development)
flask>=2.3.2,<3.0.0
fastapi>=0.97.0,<1.0.0

# Configuration management
pyyaml>=6.0,<7.0

# Testing
pytest>=7.3.1,<8.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0,<2.0.0
ipython>=8.14.0,<9.0.0

# Other utilities
tqdm>=4.65.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0